Fix package name metedata->metadata
